refactor: extract Contact into its own file
