+++ frontend/src/components/contact/ContactDetail.tsx
@@ -4,7 +4,7 @@ import { useEffect, useRef, useState } from "react";
 import EditForm from "@/components/EditForm";
 import { ContactDetailPresenter, ContactDetailView } from "@/presenter/ContactDetailPresenter";
 import { TimelineEvent } from "@/utils/mockTimelineEvents";
-import { Contact } from "@/utils/mockContacts";
+import { Contact } from "@/model/Contact";
 import { QueryState } from "@/utils/QueryState";
 
 interface Props {
