+++ frontend/src/model/Contact.ts
@@ -1,0 +1,10 @@
+export interface Contact {
+  id: number;
+  name: string;
+  phone: string;
+  email: string;
+  notes: string;
+  image?: string;
+  timeline: string[];
+  category?: string[];
+}
